+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Updated parser dependency to v0.0.10
 * Updated generation of header fields ( `name`, `version`, `variant` )
 * Updated validation of specified values in create command with new parsing functions
+* Updated compile and lint tests to use `path.join` instead of string paths
 
 ### Removed
 * Document type flag: `--documentType` (`-t`)